+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Introduced a new database table `TRIAGE` to store triage result.
+- Introduced a new database table `SEMI_MODELS` to store semi supervised models.
 - Add new `WindowsThreat` detected event for Windows Sysmon events.
 
 ## [0.20.0] - 2023-10-06